+++ app/src/main/java/tests/em_projects/com/mytestapplication/utils/DimenUtils.java
@@ -8,11 +8,11 @@ import android.content.res.Resources;
 
 public class DimenUtils {
 
-    public static int dpToPx(int dp) {
-        return (int) (dp * Resources.getSystem().getDisplayMetrics().density);
+    public static float dpToPx(int dp) {
+        return (dp * Resources.getSystem().getDisplayMetrics().density);
     }
 
-    public static int pxToDp(int px) {
-        return (int) (px / Resources.getSystem().getDisplayMetrics().density);
+    public static float pxToDp(int px) {
+        return (px / Resources.getSystem().getDisplayMetrics().density);
     }
 }
